fix: ai generated comments
