+++ sdk/src/gateway/utils.ts
@@ -1,7 +1,7 @@
 import { createPublicClient, http, zeroAddress } from 'viem';
 import { bob, bobSepolia } from 'viem/chains';
 
-import { Chain, GatewayCreateOrderRequest, OfframpOrderStatus } from './types';
+import { Chain, GatewayCreateOrderRequest, OfframpOrderStatus, OrderDetails, OrderDetailsRaw } from './types';
 import { AbiCoder, ethers } from 'ethers';
 import * as bitcoin from 'bitcoinjs-lib';
 
@@ -58,4 +58,42 @@ export function parseOrderStatus(value: number): OfframpOrderStatus {
 export function viemClient(chain: Chain) {
     const chainConfig = chain === Chain.BOB ? bob : bobSepolia;
     return createPublicClient({ chain: chainConfig, transport: http() });
+}
+
+function parseU256(value: string | null): bigint | null {
+    if (value === null) return null;
+    return BigInt(value);
+}
+
+export function convertOrderDetailsRawToOrderDetails(order: OrderDetailsRaw): OrderDetails {
+    return {
+        version: order.version,
+        data: {
+            ethAmountToReceive: parseU256(order.data.ethAmountToReceive)!,
+            maxSatsToSwapToEth: order.data.maxSatsToSwapToEth,
+            ethTransferGasLimit: parseU256(order.data.ethTransferGasLimit)!,
+            strategyGasLimit: parseU256(order.data.strategyGasLimit)!,
+            totalUserGasLimit: parseU256(order.data.totalUserGasLimit)!,
+            userGasPriceLimit: parseU256(order.data.userGasPriceLimit)!,
+            l1DataFee: parseU256(order.data.l1DataFee)!,
+            extraSatsFee: parseU256(order.data.extraSatsFee),
+            extraSatsFeeRecipient: order.data.extraSatsFeeRecipient,
+        },
+    };
+}
+export function convertOrderDetailsToRaw(order: OrderDetails): OrderDetailsRaw {
+    return {
+        version: order.version,
+        data: {
+            ethAmountToReceive: order.data.ethAmountToReceive.toString(), // bigint to string
+            maxSatsToSwapToEth: order.data.maxSatsToSwapToEth,
+            ethTransferGasLimit: order.data.ethTransferGasLimit.toString(),
+            strategyGasLimit: order.data.strategyGasLimit.toString(),
+            totalUserGasLimit: order.data.totalUserGasLimit.toString(),
+            userGasPriceLimit: order.data.userGasPriceLimit.toString(),
+            l1DataFee: order.data.l1DataFee.toString(),
+            extraSatsFee: order.data.extraSatsFee !== null ? order.data.extraSatsFee.toString() : null,
+            extraSatsFeeRecipient: order.data.extraSatsFeeRecipient,
+        },
+    };
 }
